+++ frontend/lost/src/App.jsx
@@ -6,30 +6,16 @@ import { initReactI18next } from 'react-i18next'
 import { QueryClient, QueryClientProvider } from 'react-query'
 import { Provider } from 'react-redux'
 import { BrowserRouter, Route, Routes } from 'react-router-dom'
-import deTranslation from './assets/locales/de'
-import enTranslation from './assets/locales/en'
+import deTranslation from './assets/locales/de.json'
+import enTranslation from './assets/locales/en.json'
 import { CenteredSpinner } from './components/CenteredSpinner'
 import { LostConfigProvider } from './contexts/LostConfigContext'
 import { API_URL } from './lost_settings'
 import './scss/style.scss'
 import { store } from './store'
 import { flatObj } from './utils'
-const queryClient = new QueryClient()
 
-// Containers
-const IS_DEV = true
-const VITE_BACKEND_PORT = 80
-let apiUrl = ''
-let socketUrl = ''
-if (IS_DEV) {
-  apiUrl = `${window.location.origin.replace(/:\d+?$/, '')}:${VITE_BACKEND_PORT}/api`
-  socketUrl = `${window.location.origin.replace(/:\d+?$/, '')}:${VITE_BACKEND_PORT}`
-} else {
-  apiUrl = `${window.location.origin}/api`
-  socketUrl = `${window.location.origin}`
-}
-window.API_URL = apiUrl
-window.SOCKET_URL = socketUrl
+const queryClient = new QueryClient()
 const resources = {
   en: {
     translation: flatObj(enTranslation),
@@ -38,6 +24,7 @@ const resources = {
     translation: flatObj(deTranslation),
   },
 }
+
 i18n
   .use(initReactI18next) // passes i18n down to react-i18next
   .init({
@@ -54,8 +41,8 @@ i18n
 const TheLayout = React.lazy(() => import('./coreui_containers/TheLayout'))
 
 // Pages
-const Login = React.lazy(() => import('./containers/Login/Login'))
-const Logout = React.lazy(() => import('./containers/Logout/Logout'))
+const Login = React.lazy(() => import('./containers/Authentication/Login'))
+const Logout = React.lazy(() => import('./containers/Authentication/Logout'))
 
 function App() {
   const sendError = async (event) => {
@@ -64,10 +51,9 @@ function App() {
       const usedBrowser = event.currentTarget.clientInformation.userAgent
       const location = event.currentTarget.location.href
       const decodedToken = jwtDecode(localStorage.getItem('token'))
-      let userId = null
-      if (decodedToken) {
-        userId = decodedToken.identity
-      }
+
+      const userId = decodedToken?.identity ?? null
+
       const errorObj = {
         error,
         usedBrowser,
@@ -81,11 +67,11 @@ function App() {
     }
   }
 
-  window.addEventListener('error', (event) => {
+  globalThis.addEventListener('error', (event) => {
     sendError(event)
   })
 
-  window.addEventListener('unhandledrejection', (event) => {
+  globalThis.addEventListener('unhandledrejection', (event) => {
     sendError(event)
   })
 
@@ -96,9 +82,9 @@ function App() {
           <Suspense fallback={<CenteredSpinner color="white" />}>
             <BrowserRouter>
               <Routes>
-                <Route path="/login" name="Login Page" element={<Login />} />
-                <Route path="/logout" name="Logout Page" element={<Logout />} />
-                <Route path="/*" name="Home" element={<TheLayout />} />
+                <Route path="/login" element={<Login />} />
+                <Route path="/logout" element={<Logout />} />
+                <Route path="/*" element={<TheLayout />} />
               </Routes>
             </BrowserRouter>
           </Suspense>